[MOD] Break down extract.py into utilities
- Implemented optimization utilities in `optimization_utils.py` for saving optimized programs and analyzing optimization reports.
- Developed scoring functions in `scoring.py` for various data types including dates, emails, URLs, phone numbers, and IDs, with detailed scoring logic and normalization.
- Introduced text processing utilities in `text_utils.py` for text normalization, number parsing, and handling boolean conversions.
